python/requirements.txt
click==7.1.2
grpcio==1.33.2
grpcio-tools==1.33.2
joblib==0.17.0
nltk==3.5
protobuf==3.14.0
regex==2020.11.13
six==1.15.0
tqdm==4.52.0
